+++ CHANGELOG.rst
@@ -18,6 +18,8 @@ Added
     - ``user_settings`` field in project plugins (#16)
     - Optional ``label`` key for settings
     - Optional "wait for element" args in UI test helpers to ease Javascript testing (#230)
+- **Sodarcache**
+    - Optional ``--project`` argument for the ``synccache`` command (#232)
 - **Userprofiles**
     - Form for displaying and updating user settings (#16)